+++ tests/files/src/0562_unpack_too_many.php
@@ -1,0 +1,12 @@
+<?php
+
+class Example {
+    public function __construct() {
+    }
+    public static function make(...$args) {
+        return new self(...$args);
+    }
+    public static function make2(...$args) {
+        return new self('extra', ...$args);
+    }
+}